+++ src/pages/Recommend/BodyPart/BodyPart.tsx
@@ -3,16 +3,13 @@ import { useNavigate } from "react-router-dom"
 
 import { useRecommendStore } from "@store/store"
 
-import Avatar from "@components/Avatar/Avatar"
-import Bottom from "@components/Bottom/Bottom"
-import RoundButton from "@components/Button/RoundButton"
-import ImgCheckBox from "@components/CheckBox/ImgCheckBox"
 import IconButton from "@components/IconButton/IconButton"
 import ProgressBar from "@components/Progressbar/ProgressBar"
-import SpeechBubble from "@components/SpeechBubble/SpeechBubble"
+
+import Footer from "@pages/Recommend/BodyPart/Footer"
+import List from "@pages/Recommend/BodyPart/List"
 
 import { useGetBodyPart } from "@hooks/query/useGetBodyPart"
-import { useScroll } from "@hooks/useScroll"
 
 import * as S from "../StyledRecommend"
 
@@ -21,10 +18,7 @@ const BodyPart = () => {
   const [selectedBodyParts, setSelectedBodyParts] = useState<number[]>([])
   const { setBodyPart } = useRecommendStore()
 
-  const isReady = selectedBodyParts.some((v) => v)
-
   const navigate = useNavigate()
-  const { isScrollTop } = useScroll()
 
   const handleBackPage = () => {
     navigate(-1)
@@ -50,7 +44,7 @@ const BodyPart = () => {
 
   return (
     <S.RecommendWrapper>
-      <S.Status $isScrollTop={isScrollTop}>
+      <S.Status $isScrollTop={true}>
         <IconButton
           icon="LeftArrowBold"
           size={30}
@@ -61,82 +55,15 @@ const BodyPart = () => {
           variant="round"
         />
       </S.Status>
-      <S.RecommendInner>
-        <S.RecommendGuideWrapper>
-          <S.RecommendGuide>
-            <Avatar />
-            <SpeechBubble>
-              <SpeechBubble.MainText>
-                어떤 부위의 운동을 추천해드릴까요?
-              </SpeechBubble.MainText>
-            </SpeechBubble>
-          </S.RecommendGuide>
-        </S.RecommendGuideWrapper>
-        <S.BodyPartWrapper>
-          <S.TabWrapper key="상체">
-            <S.TabTitle>{"상체"}</S.TabTitle>
-            <S.TabList>
-              {bodyParts
-                .slice(0, 6)
-                .map(({ koreanName, imgPath, bodyPartId }) => (
-                  <ImgCheckBox
-                    key={bodyPartId}
-                    src={
-                      imgPath
-                        ? imgPath
-                        : "https://www.chosun.com/resizer/v2/5O2JMBBB2NHEZOKWLS2AZYKPWU.jpg?auth=3bd18164c19b9c5b2d492ceec653597eb3e73630ac6a350cba4813bd585a1c21&width=616"
-                    }
-                    alt="테스트 이미지"
-                    handleToggle={() => {
-                      handleBodyPart(bodyPartId)
-                    }}
-                    isSelected={selectedBodyParts.includes(bodyPartId)}
-                    variant="small">
-                    {koreanName}
-                  </ImgCheckBox>
-                ))}
-            </S.TabList>
-          </S.TabWrapper>
-          <S.TabWrapper key="하체">
-            <S.TabTitle>{"하체"}</S.TabTitle>
-            <S.TabList>
-              {bodyParts
-                .slice(6)
-                .map(({ englishName, koreanName, imgPath, bodyPartId }) => (
-                  <ImgCheckBox
-                    key={englishName}
-                    src={
-                      imgPath
-                        ? imgPath
-                        : "https://www.chosun.com/resizer/v2/5O2JMBBB2NHEZOKWLS2AZYKPWU.jpg?auth=3bd18164c19b9c5b2d492ceec653597eb3e73630ac6a350cba4813bd585a1c21&width=616"
-                    }
-                    alt="테스트 이미지"
-                    handleToggle={() => {
-                      handleBodyPart(bodyPartId)
-                    }}
-                    isSelected={selectedBodyParts.includes(bodyPartId)}
-                    variant="small">
-                    {koreanName}
-                  </ImgCheckBox>
-                ))}
-            </S.TabList>
-          </S.TabWrapper>
-        </S.BodyPartWrapper>
-      </S.RecommendInner>
-      <Bottom flex="space-between">
-        <Bottom.Text>
-          {selectedBodyParts.length}개
-          <Bottom.SubText> 부위 선택됨</Bottom.SubText>
-        </Bottom.Text>
-        <RoundButton
-          onClick={handleNextPage}
-          variant="black"
-          size="big"
-          rightIcon="RightArrowWhite"
-          disabled={!isReady}>
-          다음
-        </RoundButton>
-      </Bottom>
+      <List
+        bodyParts={bodyParts}
+        selectedBodyParts={selectedBodyParts}
+        handleBodyPart={handleBodyPart}
+      />
+      <Footer
+        selectedBodyParts={selectedBodyParts}
+        handleNextPage={handleNextPage}
+      />
     </S.RecommendWrapper>
   )
 }
